+++ spec/models/activity_spec.rb
@@ -586,6 +586,28 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when gcrf_challenge_area is blank" do
+      subject { build(:programme_activity, parent: fund, gcrf_challenge_area: nil) }
+      let(:fund) { build(:fund_activity) }
+
+      it { is_expected.to be_valid(:gcrf_challenge_area_step) }
+      it { is_expected.to be_valid }
+
+      context "with a GCRF funded activity" do
+        let(:fund) { build(:fund_activity, :gcrf) }
+
+        it { is_expected.to be_invalid(:gcrf_challenge_area_step) }
+        it { is_expected.to be_invalid }
+      end
+
+      context "for a fund" do
+        subject { build(:fund_activity, :gcrf, gcrf_challenge_area: nil) }
+
+        it { is_expected.to be_valid(:gcrf_challenge_area_step) }
+        it { is_expected.to be_valid }
+      end
+    end
+
     context "when oda_eligibility is blank" do
       subject(:activity) { build(:activity, oda_eligibility: nil) }
       it "should not be valid" do