+++ spec/support/form_helpers.rb
@@ -49,6 +49,7 @@ module FormHelpers
     policy_marker_disaster_risk_reduction: "Not assessed",
     policy_marker_nutrition: "Not assessed",
     covid19_related: "4",
+    gcrf_challenge_area: "1",
     oda_eligibility: "Eligible",
     oda_eligibility_lead: Faker::Name.name,
     level:,
@@ -279,6 +280,13 @@ module FormHelpers
     expect(page).to have_content t("form.legend.activity.covid19_related")
     choose("activity[covid19_related]", option: covid19_related)
     click_button t("form.button.activity.submit")
+
+    if parent&.is_gcrf_funded?
+      expect(page).to have_content t("form.legend.activity.gcrf_challenge_area")
+      expect(page).to have_content t("form.hint.activity.gcrf_challenge_area")
+      choose("activity[gcrf_challenge_area]", option: gcrf_challenge_area)
+      click_button t("form.button.activity.submit")
+    end
 
     expect(page).to have_content t("form.legend.activity.oda_eligibility")
     expect(page).to have_content t("form.hint.activity.oda_eligibility")